New Skateboard Models and Config
Added a new skateboard model with updated UV data. Now the top and sides of a deck can have a texture. This is more realistic and means that the user can specify a clear grip tape in their design.

Also added some preferences code that will come in use when creating a settings area.

Started creating some configuration (XML) for boards that have been created.