Deck Configuration File
Added a deck configuration file and reader. Now the applications skateboard components (decks and designs) are completely configuration driven. This makes the code much neater in the scenes and means that new components can be added much more easily. Removed the SkateboardComponent class which quickly became apparant as the wrong way to handle the resources.