Ability to name boards and work on Select Deck width scene
I have added the ability to name a skateboard design in the view board scene (before the purchase section). This gives the board all of the components it needs to be stored in the save configuration and therefor I have also made it so that the design is added to the list of saves. You can see it appear in the open save scene.

Working on making the select deck width scene more robust with some comments and a restructure.