OpenSaveScene Changes
Opening a save now sets it to be the current save and lets the user edit it in the component selection scenes. If there is an existing current save then each component selection scene will load the part from the save rather.

On the view board scene, the save button will now check if a save with the given name already exists and require the user to enter a save that does not exist with that name.

I have also added and been workin g on a buy scene that lets the user select a save from the dropdown and will load some information on it. Currently there is limited information and the scene needs more work.